+++ backend/store/v2/interface.go
@@ -2,29 +2,32 @@ package v2
 
 import (
 	"github.com/sensu/sensu-go/backend/store"
-	"github.com/sensu/sensu-go/backend/store/wrap"
+	"github.com/sensu/sensu-go/backend/store/v2/wrap"
 )
+
+// Type alias for documentation clarity
+type Wrapper = wrap.Wrapper
 
 // Interface specifies the interface of a v2 store.
 type Interface interface {
 	// CreateOrUpdate creates or updates the wrapped resource.
-	CreateOrUpdate(ResourceRequest, *wrap.Wrapper) error
+	CreateOrUpdate(ResourceRequest, *Wrapper) error
 
 	// UpdateIfExists updates the resource with the wrapped resource, but only
 	// if it already exists in the store.
-	UpdateIfExists(ResourceRequest, *wrap.Wrapper) error
+	UpdateIfExists(ResourceRequest, *Wrapper) error
 
 	// CreateIfNotExists writes the wrapped resource to the store, but only if
 	// it does not already exist.
-	CreateIfNotExists(ResourceRequest, *wrap.Wrapper) error
+	CreateIfNotExists(ResourceRequest, *Wrapper) error
 
 	// Get gets a wrapped resource from the store.
-	Get(ResourceRequest) (*wrap.Wrapper, error)
+	Get(ResourceRequest) (*Wrapper, error)
 
 	// Delete deletes a resource from the store.
 	Delete(ResourceRequest) error
 
 	// List lists all resources specified by the resource request, and the
 	// selection predicate.
-	List(ResourceRequest, *store.SelectionPredicate) ([]*wrap.Wrapper, error)
+	List(ResourceRequest, *store.SelectionPredicate) ([]*Wrapper, error)
 }
